Move to webhook.ts types file
