+++ src/types/types.ts
@@ -12,5 +12,4 @@ export * from "./options.ts";
 export * from "./permission.ts";
 export * from "./presence.ts";
 export * from "./role.ts";
-export * from "./slash.ts";
 export * from "./webhook.ts";
